09 start adding in tests for refs and testing for structure as well as content
